+++ engineioxide/src/transport/ws.rs
@@ -58,16 +58,9 @@ pub fn new_req<R: Send + 'static, B, H: EngineIoHandler>(
     sid: Option<Sid>,
     req: Request<R>,
 ) -> Result<Response<ResponseBody<B>>, Error> {
-    let mut parts = Request::builder()
-        .method(req.method().clone())
-        .uri(req.uri().clone())
-        .version(req.version())
-        .body(())
-        .unwrap()
-        .into_parts()
-        .0;
+    let (parts, body) = req.into_parts();
+    let req = Request::from_parts(parts.clone(), body);
 
-    parts.headers.extend(req.headers().clone());
     let ws_key = parts
         .headers
         .get("Sec-WebSocket-Key")
@@ -121,7 +114,7 @@ where
     let (socket, ws) = if let Some(sid) = sid {
         match engine.get_socket(sid) {
             None => return Err(Error::UnknownSessionID(sid)),
-            Some(socket) if socket.is_ws() => return Err(Error::UpgradeError),
+            Some(socket) if socket.is_ws() => return Err(Error::Upgrade),
             Some(socket) => {
                 let mut ws = ws_init().await;
                 upgrade_handshake::<H, S>(&socket, &mut ws).await?;
@@ -306,7 +299,7 @@ where
     // Fetch the next packet from the ws stream, it should be a PingUpgrade packet
     let msg = match ws.next().await {
         Some(Ok(Message::Text(d))) => d,
-        _ => Err(Error::UpgradeError)?,
+        _ => Err(Error::Upgrade)?,
     };
     match Packet::try_from(msg)? {
         Packet::PingUpgrade => {
@@ -326,12 +319,12 @@ where
         Some(Ok(Message::Close(_))) => {
             #[cfg(feature = "tracing")]
             tracing::debug!("ws stream closed before upgrade");
-            Err(Error::UpgradeError)?
+            Err(Error::Upgrade)?
         }
         _ => {
             #[cfg(feature = "tracing")]
             tracing::debug!("unexpected ws message before upgrade");
-            Err(Error::UpgradeError)?
+            Err(Error::Upgrade)?
         }
     };
     match Packet::try_from(msg)? {