+++ socketioxide/src/handler/extract.rs
@@ -13,7 +13,7 @@
 //! * [`AckSender`]: Can be used to send an ack response to the current message event
 //! * [`ProtocolVersion`](crate::ProtocolVersion): extracts the protocol version
 //! * [`TransportType`](crate::TransportType): extracts the transport type
-//! * [`DisconnectReason`](crate::socket::DisconnectReason): extracts the reason of the disconnection
+//! * [`DisconnectReason`]: extracts the reason of the disconnection
 //! * [`State`]: extracts a reference to a state previously set with [`SocketIoBuilder::with_state`](crate::io::SocketIoBuilder).
 //!
 //! ### You can also implement your own Extractor with the [`FromConnectParts`], [`FromMessageParts`] and [`FromDisconnectParts`] traits
@@ -88,12 +88,12 @@ use std::sync::Arc;
 use super::message::FromMessageParts;
 use super::FromDisconnectParts;
 use super::{connect::FromConnectParts, message::FromMessage};
+use crate::errors::{DisconnectError, SendError};
 use crate::socket::DisconnectReason;
 use crate::{
     adapter::{Adapter, LocalAdapter},
     packet::Packet,
     socket::Socket,
-    SendError,
 };
 use serde::{de::DeserializeOwned, Serialize};
 use serde_json::Value;
@@ -180,6 +180,7 @@ where
     }
 }
 /// An Extractor that returns a reference to a [`Socket`].
+#[derive(Debug)]
 pub struct SocketRef<A: Adapter = LocalAdapter>(Arc<Socket<A>>);
 
 impl<A: Adapter> FromConnectParts<A> for SocketRef<A> {
@@ -213,18 +214,25 @@ impl<A: Adapter> std::ops::Deref for SocketRef<A> {
         &self.0
     }
 }
-
-impl<A: Adapter> SocketRef<A> {
+impl<A: Adapter> PartialEq for SocketRef<A> {
     #[inline(always)]
-    pub(crate) fn new(socket: Arc<Socket<A>>) -> Self {
+    fn eq(&self, other: &Self) -> bool {
+        self.0.id == other.0.id
+    }
+}
+impl<A: Adapter> From<Arc<Socket<A>>> for SocketRef<A> {
+    #[inline(always)]
+    fn from(socket: Arc<Socket<A>>) -> Self {
         Self(socket)
     }
+}
 
+impl<A: Adapter> SocketRef<A> {
     /// Disconnect the socket from the current namespace,
     ///
     /// It will also call the disconnect handler if it is set.
     #[inline(always)]
-    pub fn disconnect(self) -> Result<(), SendError> {
+    pub fn disconnect(self) -> Result<(), DisconnectError> {
         self.0.disconnect()
     }
 }
@@ -288,7 +296,7 @@ impl<A: Adapter> AckSender<A> {
             } else {
                 Packet::bin_ack(ns, data, self.binary, ack_id)
             };
-            self.socket.send(packet)
+            Ok(self.socket.send(packet)?)
         } else {
             Ok(())
         }